refactor: consolidate button components and add icon-slate variant
- Remove redundant ButtonVariants.tsx and Buttons.tsx files
- Update imports to use unified button.tsx component
- Add new 'icon-slate' variant for icon buttons with dark slate background
- Add 'xs' size variant for smaller buttons
- Update button styling to use pg1-bold font class
- Apply new icon-slate variant to Footer contact button
- Minor code cleanup and formatting fixes
