+++ team-mars-frontend/src/components/ui/button.tsx
@@ -5,7 +5,7 @@ import { cva, type VariantProps } from 'class-variance-authority';
 import { cn } from '@/lib/utils';
 
 const buttonVariants = cva(
-  "inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-[2px] text-sm font-medium transition-all disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4 shrink-0 [&_svg]:shrink-0 outline-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive",
+  "inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-[2px] pg1-bold transition-all disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4 shrink-0 [&_svg]:shrink-0 outline-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive",
   {
     variants: {
       variant: {
@@ -22,14 +22,16 @@ const buttonVariants = cva(
           'border border-secondary text-secondary-foreground bg-transparent hover:bg-secondary hover:text-secondary-foreground shadow-xs',
         'nav-primary':
           'border-2 border-primary text-gray-600 bg-transparent hover:bg-transparent hover:text-foreground shadow-xs group cursor-pointer [&_svg]:transition-transform',
-        secondary: 'bg-secondary text-secondary-foreground hover:bg-secondary/80',
+        secondary: 'bg-secondary text-secondary-foreground hover:bg-secondary/80 cursor-pointer',
+        'icon-slate': 'bg-slate-600 border-2 border-white text-white hover:bg-white hover:text-primary rounded-[2px] shadow-sm',
         ghost: 'hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50',
         link: 'text-primary underline-offset-4 hover:underline',
       },
       size: {
-        default: 'h-9 px-4 py-2 has-[>svg]:px-3',
-        sm: 'h-8 rounded-[2px] gap-1.5 px-3 has-[>svg]:px-2.5',
-        lg: 'h-10 rounded-[2px] px-6 has-[>svg]:px-4',
+        default: 'h-auto px-4 py-2 has-[>svg]:px-3',
+        xs: 'h-auto rounded-[2px] gap-1 px-[12px] py-[6px] has-[>svg]:px-2',
+        sm: 'h-auto rounded-[2px] gap-1.5 px-[16px] py-[8px] has-[>svg]:px-2.5',
+        lg: 'h-auto rounded-[2px] gap-2 px-[24px] py-[12px] has-[>svg]:px-4',
         icon: 'size-9',
         'icon-sm': 'size-8',
         'icon-lg': 'size-10',
